requirements.txt
python-dateutil>=2.9.0
rich>=13.7.1
